+++ src/lib.rs
@@ -3,5 +3,6 @@ pub mod diagnostic;
 mod javadoc;
 pub mod parser;
 mod rules;
+mod validation;
 
 pub use parser::{ParseFileResult, Parser};