+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+- Fix wrong types for maps
+- Validate method arg direction based on type
+- General validation and diagnostics improvements
+
 ## [0.1.5] - 2021-12-22
 
 - Initial version